py/requirements.txt
anthropic>=0.40.0
firecrawl>=1.0.0
openai>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
